Fix bug with protected ranges (#13)
When specifying the range to protect, the default editors seems to be
anyone with edit access to the spreadsheet. Since we are sharing the
spreadsheets with volunteers and giving them edit access, protecting the
ranges essentially has no effect, since they can still edit them. Now,
we are using the owner of the spreadsheet as the single editor of the
protected range.